comment authorization

+++ BACKEND/routes/comment.js
@@ -16,17 +16,28 @@ router.post("/:id", (req, res, next) => {
 //update routes
 router.patch("/:id", (req, res, next) => {
   let id = req.params.id;
-  Comment.findOneAndUpdate(id, req.body, (err, updatedPost) => {
+  Comment.findById(id, (err, comment) => {
     if (err) return next(err);
-    res.json({ status: "sucess", message: "Comment updated" });
+    if (req.userid == comment.userid) {
+      Comment.findOneAndUpdate(id, req.body, (err, updatedPost) => {
+        if (err) return next(err);
+        res.status(200).json({ status: "sucess", message: "Comment updated" });
+      });
+    } else res.status(401).json({ status: "fail", message: "Not authorised!" });
   });
 });
+
 //delete Article and comments
 router.delete("/:id", (req, res, next) => {
   let id = req.params.id;
-  Comment.findOneAndDelete(id, (err, msg) => {
+  Comment.findById(id, (err, comment) => {
     if (err) return next(err);
-    Article.findByIdAndUpdate(msg.post, { $pull: { comments: id } });
+    if (req.userid == comment.userid) {
+      Comment.findOneAndDelete(id, (err, msg) => {
+        if (err) return next(err);
+        Article.findByIdAndUpdate(msg.post, { $pull: { comments: id } });
+      });
+    } else res.status(401).json({ status: "fail", message: "Not authorised!" });
   });
 });
